refactor (typing): move SortMap into Disco.Types
Move the definition of SortMap from Disco.Types.Rules into Disco.Types,
in preparation for adding SortMap to PolyType (allowing qualified types
to show up in surface syntax).

See #179.
